+++ packages/kernel-adapter-postgres/src/postgres/index.ts
@@ -1,3 +1,4 @@
 export { PostgresMetaKernelRepository } from "./repository/postgres-meta-kernel.repository";
 export { createPostgresMetaKernelRepository } from "./factory/postgres-meta-kernel-repository.factory";
+export { PostgresMetaKernelRepositoryFactory } from "./factory/postgres-meta-kernel-repository-factory.impl";
 export type { MigrationExecutionContext } from "./interfaces";