+++ packages/kernel/src/core/interfaces/index.ts
@@ -1,5 +1,6 @@
 export * from "./kernel.interface";
 export * from "./in-memory-meta-definition-seed.interface";
 export * from "./meta-kernel.interface";
+export * from "./meta-kernel-repository-factory.interface";
 export * from "./migration-safety.interface";
 export * from "./schema-diff.interface";